move jean cons to cons.go
